ch1/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==5.0.1
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.47
langchain-groq==0.3.1
langchain-text-splitters==0.3.7
langsmith==0.3.18
orjson==3.10.15
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zstandard==0.23.0
